앱바 설정 아이콘 추가, 바텀 네게이션추가
