+++ lib/components/custom_app_bar.dart
@@ -1,21 +1,42 @@
+import 'package:final_project_team02/holder/profile/settings_page.dart';
 import 'package:flutter/material.dart';
 
-import '../holder/welcome/welcome_page.dart';
+import '../holder/main_hoder.dart';
 
 class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
-  final IconData micon;
+  final IconData? miconRight;
+  final IconData? miconLeft;
 
-  const CustomAppBar({required this.micon});
+  const CustomAppBar({
+    required this.miconRight,
+    required this.miconLeft,
+  });
 
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      title: TextButton(
-        onPressed: () {
-          Navigator.pop(
-              context, MaterialPageRoute(builder: (context) => WelcomePage()));
-        },
-        child: Icon(micon),
+      title: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          TextButton(
+            onPressed: () {
+              Navigator.pushAndRemoveUntil(
+                context,
+                MaterialPageRoute(builder: (context) => MainHolder()),
+                ModalRoute.withName(
+                    '/home'), // '/firstPage' 라우트를 제외한 모든 페이지를 스택에서 제거
+              );
+            },
+            child: Icon(miconRight),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.push(context,
+                  MaterialPageRoute(builder: (context) => SettingsPage()));
+            },
+            child: Icon(miconLeft),
+          ),
+        ],
       ),
     );
   }
