+++ lib/holder/MyPage/creator/components/_creator_page_body.dart
@@ -2,57 +2,33 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-class MyPageBody extends StatelessWidget {
-  const MyPageBody({super.key});
-
+class CreatorPageBody extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Stack(
+    return Column(
       children: [
-        Column(
-          children: [
-            const SizedBox(height: 20),
-            Expanded(
-              child: DefaultTabController(
-                length: 2,
-                child: Column(
-                  children: [
-                    //탭바
-                    _buildTabBar(),
-                    Expanded(
-                      child: TabBarView(
-                        children: [
-                          //탭바 그리드
-                          _buildTabGrid(),
-                          //탭바 옷장
-                          _buildTabCloset(),
-                        ],
-                      ),
-                    ),
-                  ],
+        const SizedBox(height: 20),
+        Expanded(
+          child: DefaultTabController(
+            length: 2,
+            child: Column(
+              children: [
+                //탭바
+                _buildTabBar(),
+                Expanded(
+                  child: TabBarView(
+                    children: [
+                      //탭바 그리드
+                      _buildTabGrid(),
+                      //탭바 옷장
+                      _buildTabCloset(),
+                    ],
+                  ),
                 ),
-              ),
+              ],
             ),
-          ],
+          ),
         ),
-        Positioned(
-            bottom: 66,
-            left: 10,
-            right: 10,
-            child: Container(
-              width: 350,
-              decoration: BoxDecoration(
-                color: Colors.black,
-                borderRadius: BorderRadius.circular(10),
-              ),
-              child: TextButton(
-                onPressed: () {},
-                child: Text(
-                  "지원하기",
-                  style: TextStyle(color: Colors.white, fontSize: 30),
-                ),
-              ),
-            ))
       ],
     );
   }
@@ -90,7 +66,7 @@ class MyPageBody extends StatelessWidget {
             children: [
               Expanded(
                 child: AspectRatio(
-                  aspectRatio: 4,
+                  aspectRatio: 2 / 1,
                   child: ClipRRect(
                     child: Image.network(
                         // "https://picsum.photos/400/400",
@@ -109,12 +85,10 @@ class MyPageBody extends StatelessWidget {
                             color: Colors.black,
                             fontSize: 25,
                             fontWeight: FontWeight.bold)),
-                    Text(
-                      "옷설명",
-                      style: TextStyle(
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
+                    Text("옷설명",
+                        style: TextStyle(
+                          overflow: TextOverflow.ellipsis,
+                        )),
                     //numberFormat 사용하면되
                     Text(
                       "가격 : 10,000원",
@@ -123,7 +97,7 @@ class MyPageBody extends StatelessWidget {
                     SizedBox(height: 15),
                   ],
                 ),
-              ),
+              )
             ],
           ),
         );
@@ -132,7 +106,42 @@ class MyPageBody extends StatelessWidget {
   }
 
   Widget _buildTabGrid() {
-    return Align(child: Text("크리에이터가 되어 자신의 멋진 코디를 뽐내 보세요.!!"));
+    return GridView.builder(
+      physics: NeverScrollableScrollPhysics(),
+      shrinkWrap: true,
+      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+        crossAxisSpacing: 2,
+        crossAxisCount: 2,
+        mainAxisSpacing: 2,
+      ),
+      itemCount: 42,
+      itemBuilder: (context, index) {
+        return Stack(
+          children: [
+            InkWell(
+              onTap: () {},
+              child: AspectRatio(
+                aspectRatio: 1,
+                child: Image.network(
+                    "https://picsum.photos/id/${index + 1}/1000/1000",
+                    fit: BoxFit.cover),
+              ),
+            ),
+            Positioned(
+              right: 10,
+              top: 10,
+              child: Icon(
+                Icons.auto_awesome_motion,
+                color: Colors.white,
+              ),
+              //auto_awesome_motion
+              //style
+              //filter_none
+            )
+          ],
+        );
+      },
+    );
   }
 
   //가격 1,000원 쉼표, 넣기
