+++ lib/holder/MyPage/creator/components/_creator_page_header.dart
@@ -1,18 +1,14 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:intl/intl.dart';
 
-import '../../../components/custom_button.dart';
-import '../../../theme.dart';
+import '../../../../theme.dart';
 
 class CreatorPageheader extends StatelessWidget {
-  const CreatorPageheader({super.key});
-
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20),
+      padding: EdgeInsets.symmetric(horizontal: 20),
       child: Column(
         children: [
           SizedBox(height: 30),
@@ -20,7 +16,6 @@ class CreatorPageheader extends StatelessWidget {
           SizedBox(height: 20),
           _buildCreatorComment(),
           SizedBox(height: 20),
-          custom_button(title: "팔로우"),
         ],
       ),
     );
@@ -67,4 +62,3 @@ class CreatorPageheader extends StatelessWidget {
     );
   }
 }
-
