+++ lib/holder/MyPage/user/components/user_my_page_body.dart
@@ -2,35 +2,55 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-class CreatorPageBody extends StatelessWidget {
-  const CreatorPageBody({super.key});
-
+class MyPageBody extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Column(
+    return Stack(
       children: [
-        const SizedBox(height: 20),
-        Expanded(
-          child: DefaultTabController(
-            length: 2,
-            child: Column(
-              children: [
-                //탭바
-                _buildTabBar(),
-                Expanded(
-                  child: TabBarView(
-                    children: [
-                      //탭바 그리드
-                      _buildTabGrid(),
-                      //탭바 옷장
-                      _buildTabCloset(),
-                    ],
-                  ),
+        Column(
+          children: [
+            SizedBox(height: 20),
+            Expanded(
+              child: DefaultTabController(
+                length: 2,
+                child: Column(
+                  children: [
+                    //탭바
+                    _buildTabBar(),
+                    Expanded(
+                      child: TabBarView(
+                        children: [
+                          //탭바 그리드
+                          _buildTabGrid(),
+                          //탭바 옷장
+                          _buildTabCloset(),
+                        ],
+                      ),
+                    ),
+                  ],
                 ),
-              ],
+              ),
             ),
-          ),
+          ],
         ),
+        Positioned(
+            bottom: 66,
+            left: 10,
+            right: 10,
+            child: Container(
+              width: 350,
+              decoration: BoxDecoration(
+                color: Colors.black,
+                borderRadius: BorderRadius.circular(10),
+              ),
+              child: TextButton(
+                onPressed: () {},
+                child: Text(
+                  "지원하기",
+                  style: TextStyle(color: Colors.white, fontSize: 30),
+                ),
+              ),
+            ))
       ],
     );
   }
@@ -78,7 +98,7 @@ class CreatorPageBody extends StatelessWidget {
                 ),
               ),
               Padding(
-                padding: const EdgeInsets.only(left: 10),
+                padding: EdgeInsets.only(left: 10),
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
@@ -101,7 +121,7 @@ class CreatorPageBody extends StatelessWidget {
                     SizedBox(height: 15),
                   ],
                 ),
-              )
+              ),
             ],
           ),
         );
@@ -110,42 +130,7 @@ class CreatorPageBody extends StatelessWidget {
   }
 
   Widget _buildTabGrid() {
-    return GridView.builder(
-      physics: NeverScrollableScrollPhysics(),
-      shrinkWrap: true,
-      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisSpacing: 2,
-        crossAxisCount: 2,
-        mainAxisSpacing: 2,
-      ),
-      itemCount: 42,
-      itemBuilder: (context, index) {
-        return Stack(
-          children: [
-            InkWell(
-              onTap: () {},
-              child: AspectRatio(
-                aspectRatio: 1,
-                child: Image.network(
-                    "https://picsum.photos/id/${index + 1}/600/600",
-                    fit: BoxFit.cover),
-              ),
-            ),
-            Positioned(
-              right: 10,
-              top: 10,
-              child: Icon(
-                Icons.auto_awesome_motion,
-                color: Colors.white,
-              ),
-              //auto_awesome_motion
-              //style
-              //filter_none
-            )
-          ],
-        );
-      },
-    );
+    return Align(child: Text("크리에이터가 되어 자신의 멋진 코디를 뽐내 보세요.!!"));
   }
 
   //가격 1,000원 쉼표, 넣기
