+++ lib/holder/MyPage/user/components/user_my_page_header.dart
@@ -5,8 +5,6 @@ import '../../../../components/custom_button.dart';
 import '../../../../theme.dart';
 
 class MyPageheader extends StatelessWidget {
-  const MyPageheader({super.key});
-
   @override
   Widget build(BuildContext context) {
     return Padding(
