+++ lib/holder/MyPage/user/user_my_page.dart
@@ -1,16 +1,14 @@
 import 'package:flutter/material.dart';
 
 import '../../../components/custom_app_bar.dart';
-import 'components/_my_page_body.dart';
-import 'components/_my_page_header.dart';
+import 'components/user_my_page_body.dart';
+import 'components/user_my_page_header.dart';
 
-class MyPage extends StatelessWidget {
-  const MyPage({super.key});
-
+class UserMyPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: CustomAppBar(micon: Icons.home),
+      appBar: CustomAppBar(miconRight: Icons.home, miconLeft: Icons.settings),
       backgroundColor: Colors.white,
       body: NestedScrollView(
         headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
