+++ lib/holder/main/componenets/_auto_ad_slider.dart
